src/requirements.txt
requests
numpy
pyjwt
fastapi[standard]
passlib[bcrypt]

SQLAlchemy
GeoAlchemy2
psycopg[binary]
scipy
geopy

# optional:
matplotlib